tests-playwright/requirements.txt
pytest == 8.3.*
pytest-playwright == 0.6.*
